Break parse.go into record.go & tree.go
