+++ pkg/lex/tree_test.go
